trainer/requirements.txt
opencv-python==4.5.3.56
Pillow==9.3.0
Polygon3==3.0.9.1
PyYAML==5.4.1
scikit-image==0.17.2
Shapely==1.8.0
torch==1.7.1
torchvision==0.10.0
wandb==0.12.9
numpy==1.23.1
tqdm  
matplotlib
pandas
natsort
nltk
mlflow